requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uuid==1.30
flask-cors==4.0.1
flask-jwt-extended==4.6.0
gunicorn==23.0.0
sqlalchemy==2.0.31
Flask-SQLAlchemy==3.1.1
requests==2.32.3
boto3==1.35.3
stripe==10.8.0
gevent==24.2.1
flask-bcrypt==1.0.1